nghttpx: Rework sending and receiving ECN bits

+++ src/shrpx_quic_listener.cc
@@ -59,8 +59,7 @@ void QUICListener::on_read() {
   msg.msg_iov = &msg_iov;
   msg.msg_iovlen = 1;
 
-  uint8_t msg_ctrl[CMSG_SPACE(sizeof(uint8_t)) +
-                   CMSG_SPACE(sizeof(in6_pktinfo)) +
+  uint8_t msg_ctrl[CMSG_SPACE(sizeof(int)) + CMSG_SPACE(sizeof(in6_pktinfo)) +
                    CMSG_SPACE(sizeof(uint16_t))];
   msg.msg_control = msg_ctrl;
